support in rebuild_all.sh for architecture-specific packages

+++ rebuild_all.sh
@@ -6,16 +6,25 @@ set -x
 cd $(dirname $0)
 
 for pkg in $(cat packages.order); do
-    [ -e rpmbuild/SRPMS/"$(/usr/local/bin/rpmspec -q --srpm rpmbuild/SPECS/$pkg.spec | sed -E 's|\.[^.]*$||')".src.rpm ] \
-        || {
-            ./rebuild.sh rpmbuild/SPECS/$pkg.spec
-            cd repo
-            rm -rf *
-            cp -r ../rpmbuild/RPMS/* .
-            createrepo_c .
-            dnf makecache --repo local
-            cd ..
-        }
+    if [[ $pkg = *":"* ]]; then
+        # architecture-specific package
+        arch=${pkg%:*}
+        pkg=${pkg#*:}
+    else
+        arch=""
+    fi
+    if [[ -z $arch ]] || [[ $arch = $(uname -m) ]]; then
+        [ -e rpmbuild/SRPMS/"$(/usr/local/bin/rpmspec -q --srpm rpmbuild/SPECS/$pkg.spec | sed -E 's|\.[^.]*$||')".src.rpm ] \
+            || {
+                ./rebuild.sh rpmbuild/SPECS/$pkg.spec
+                cd repo
+                rm -rf *
+                cp -r ../rpmbuild/RPMS/* .
+                createrepo_c .
+                dnf makecache --repo local
+                cd ..
+            }
+    fi
 done
 
 if [ $(uname -m) = "arm64" ]; then
